moved refinements into extensions module, removed tty color support from #log
